+++ NEWS.md
@@ -1,3 +1,10 @@
-## pivotaltrackR 0.1.0
+# pivotaltrackR 0.1.1 (under development)
+
+* `labels()` to retrieve the names of labels associated with a story
+* `addLabel()`/`addLabels()` to add labels to a story
+* Better print method for `stories`
+
+
+# pivotaltrackR 0.1.0
 
 * Query Pivotal "Stories" with `getStories()`. To deal with individual stories, `createStory()`, `getStory()`, `editStory()`, and `deleteStory()`.